[1.2.0-original] Fix: Importing in admin within a transaction gives TransactionManagementError.
If we are using MySQL as database, the django.contrib.admin app, and
some other configurations that run the admin views inside an atomic
block (a transaction), when importing data we now test if we are inside
an atomic block to know if the DB supports transactions, in addition to
previous check to see if the DB features supports transactions.

This is to solve issue #609.

# Conflicts:
#	AUTHORS